Frame renamed to StackElement for implicit search
Frame allowed for an implicit java.lang.StackTraceElement
parameter.  Convenient for the method creator, but not
for the caller, since there is no guidance provided to the
user or compiler on how the implicit parameter should be
provided.

Instead, a new StackElement class is the implicit
parameter type.  At the site of the callee,
the compiler automatically finds the implicit
by searching StackElement's companion object's implicit
definitions when the method is being called, even
if the callee has not imported StackElement.